requirements.txt
Flask==2.3.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.6
Jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0

# Image and file handling
Pillow==9.5.0

# Frontend enhancements (optional, if using local bundling)
Flask-Bootstrap==3.3.7.1

# Geo functionality (optional, depending on server-side validation)
geopy==2.3.0

# If you use a PDF writer/generator in future (like for WIL logs or printable reports)
reportlab==4.0.5

# For future REST API support (optional)
Flask-RESTful==0.3.9
